feat(workspace-cap): port search leaf (slice 4/N)
Verbatim git mv of search.rs (937 lines, last true leaf in workspace/)
into crates/dasclaw_workspace_cap/src/. Old path kept as pub-use
re-export shim per ADR-129 §1.3 + ADR-152 §3 F3.4.

External deps only: std::collections::HashMap, uuid.
No internal crate::workspace / crate::llm / crate::error deps.

lib.rs: pub mod search;
Cargo.toml: no new deps (uuid already added in slice 3).

Remaining 5 files (embeddings/embedding_cache/hygiene/repository/mod.rs)
are not leaves and need llm/error/bootstrap/safety crate-ification
first (phase 2 of ADR-152).

Local validation: cargo nextest 113 passed; dasclaw check pass;
fmt/no_panics/clippy clean.

Closes #697
Refs #674
