+++ crates/dasclaw_workspace_cap/src/lib.rs
@@ -50,6 +50,7 @@ pub mod document;
 pub mod layer;
 pub mod policy;
 pub mod privacy;
+pub mod search;
 pub use policy::{
     NetworkAccess, SandboxPolicy, WritableRoot, default_read_only_subpaths_for_writable_root,
 };
